### 2020-12-17 #2918 @whedon
[PRE REVIEW]: NiaAML: AutoML framework based on stochastic population-based nature-inspired algorithms
**Submitting author:** @lukapecnik (<a href="http://orcid.org/0000-0002-3897-9774">Luka Pečnik</a>)
**Repository:** <a href="https://github.com/lukapecnik/NiaAML" target ="_blank">https://github.com/lukapecnik/NiaAML</a>
**Version:** 1.1.0
