### 2020-12-18 2c88056
updated examples for a quicker execution [ci skip]
