### 2020-12-19 #2918 @danielskatz
Thanks @adi3 - once there is an editor assigned, they will decide on reviewers, and likely accept your offer
