### 2020-12-21 c46948f
original method's optimization support
