+++ niaaml/pipeline_optimizer.py
@@ -1,5 +1,5 @@
 import numpy as np
-from niaaml.pipeline import Pipeline
+from niaaml.pipeline import Pipeline, _PipelineBenchmark
 from niaaml.classifiers import ClassifierFactory
 from niaaml.preprocessing.feature_selection import FeatureSelectionAlgorithmFactory
 from niaaml.preprocessing.feature_transform import FeatureTransformAlgorithmFactory
@@ -9,11 +9,15 @@ from NiaPy.algorithms.utility import AlgorithmUtility
 from niaaml.utilities import get_bin_index
 from niaaml.preprocessing.encoding.utility import encode_categorical_features
 from niaaml.preprocessing.imputation.utility import impute_features
+from niaaml.fitness import FitnessFactory
 from niaaml.logger import Logger
 import pandas as pd
 
 __all__ = [
-    'PipelineOptimizer'
+    'PipelineOptimizer',
+    '_PipelineOptimizerBenchmark',
+    '_PipelineOptimizerBenchmarkV1',
+    '_PipelineOptimizerBenchmarkV2'
 ]
 
 class PipelineOptimizer:
@@ -158,10 +162,72 @@ class PipelineOptimizer:
 
         self.__data.set_x(features)
 
-        benchmark = _PipelineOptimizerBenchmark(self, fitness_name, inner_population_size, number_of_inner_evaluations, inner_optimization_algorithm if inner_optimization_algorithm is not None else optimization_algorithm)
+        benchmark = _PipelineOptimizerBenchmarkV2(self, fitness_name, inner_population_size, number_of_inner_evaluations, inner_optimization_algorithm if inner_optimization_algorithm is not None else optimization_algorithm)
         task = StoppingTask(
             D=3,
             nFES=number_of_pipeline_evaluations,
+            benchmark=benchmark
+            )
+        algo.run(task)
+        
+        pipeline = benchmark.get_pipeline()
+        if pipeline is not None:
+            pipeline.set_categorical_features_encoders(self.__categorical_features_encoders)
+            pipeline.set_imputers(self.__imputers)
+
+        return pipeline
+
+    def run_v1(self, fitness_name, population_size, number_of_evaluations, optimization_algorithm):
+        r"""Run classification pipeline optimization process according to the original NiaAML paper.
+
+        Reference:
+            Fister, Iztok, Milan Zorman, and Dušan Fister. "Continuous Optimizers for Automatic Design and Evaluation of Classification Pipelines." Frontier Applications of Nature Inspired Computation. Springer, Singapore, 2020. 281-301.
+
+        Arguments:
+            fitness_name (str): Name of the fitness class to use as a function.
+            population_size (uint): Number of individuals in the optimization process.
+            number_of_evaluations (uint): Number of maximum evaluations.
+            optimization_algorithm (str): Name of the optimization algorithm to use.
+        
+        Returns:
+            Pipeline: Best pipeline found in the optimization process.
+        """
+
+        algo = self.__niapy_algorithm_utility.get_algorithm(optimization_algorithm)
+        algo.NP = population_size
+
+        features = self.__data.get_x()
+
+        if self.__imputer is not None:
+            features, self.__imputers = impute_features(features, self.__imputer)
+
+        if self.__categorical_features_encoder is not None:
+            features, self.__categorical_features_encoders = encode_categorical_features(features, self.__categorical_features_encoder)
+
+        self.__data.set_x(features)
+
+        D = 3
+        factories = [
+            (self.__feature_selection_algorithms, FeatureSelectionAlgorithmFactory()),
+            (self.__feature_transform_algorithms, FeatureTransformAlgorithmFactory()),
+            (self.__classifiers, ClassifierFactory())
+            ]
+
[…]
