### 2020-12-22 76428ce
examples, readme, tests and docs updated
