+++ changelog.rst
@@ -1,192 +1,0 @@
-Changelog
-=========
-
-`1.0.0rc4 <https://github.com/lukapecnik/niaaml/tree/1.0.0rc4>`__ (2020-12-10)
-------------------------------------------------------------------------------
-
-`Full
-Changelog <https://github.com/lukapecnik/niaaml/compare/1.0.0rc3...1.0.0rc4>`__
-
-**Merged pull requests:**
-
--  Python 3.6 support
-   `#33 <https://github.com/lukapecnik/NiaAML/pull/33>`__
-   (`lukapecnik <https://github.com/lukapecnik>`__)
--  First version of paper
-   `#32 <https://github.com/lukapecnik/NiaAML/pull/32>`__
-   (`firefly-cpp <https://github.com/firefly-cpp>`__)
-
-`1.0.0rc3 <https://github.com/lukapecnik/niaaml/tree/1.0.0rc3>`__ (2020-12-10)
-------------------------------------------------------------------------------
-
-`Full
-Changelog <https://github.com/lukapecnik/niaaml/compare/1.0.0rc2...1.0.0rc3>`__
-
-`1.0.0rc2 <https://github.com/lukapecnik/niaaml/tree/1.0.0rc2>`__ (2020-12-08)
-------------------------------------------------------------------------------
-
-`Full
-Changelog <https://github.com/lukapecnik/niaaml/compare/1.0.0rc1...1.0.0rc2>`__
-
-**Merged pull requests:**
-
--  feature missing values imputation
-   `#30 <https://github.com/lukapecnik/NiaAML/pull/30>`__
-   (`lukapecnik <https://github.com/lukapecnik>`__)
--  README updated `#29 <https://github.com/lukapecnik/NiaAML/pull/29>`__
-   (`firefly-cpp <https://github.com/firefly-cpp>`__)
--  Readme update, encoder check for type int64
-   `#28 <https://github.com/lukapecnik/NiaAML/pull/28>`__
-   (`lukapecnik <https://github.com/lukapecnik>`__)
--  Markdown and docs
-   `#27 <https://github.com/lukapecnik/NiaAML/pull/27>`__
-   (`lukapecnik <https://github.com/lukapecnik>`__)
-
-`1.0.0rc1 <https://github.com/lukapecnik/niaaml/tree/1.0.0rc1>`__ (2020-12-06)
-------------------------------------------------------------------------------
-
-`Full
-Changelog <https://github.com/lukapecnik/niaaml/compare/0.1.4...1.0.0rc1>`__
-
-**Merged pull requests:**
-
--  Added support for categorical features
-   `#26 <https://github.com/lukapecnik/NiaAML/pull/26>`__
-   (`lukapecnik <https://github.com/lukapecnik>`__)
-
-`0.1.4 <https://github.com/lukapecnik/niaaml/tree/0.1.4>`__ (2020-12-05)
-------------------------------------------------------------------------
-
-`Full
-Changelog <https://github.com/lukapecnik/niaaml/compare/0.1.3...0.1.4>`__
-
-**Merged pull requests:**
-
--  remove 10-fold cross validation from benchmark, critical bug fix
-   `#25 <https://github.com/lukapecnik/NiaAML/pull/25>`__
-   (`lukapecnik <https://github.com/lukapecnik>`__)
--  run all experiments at once
-   `#24 <https://github.com/lukapecnik/NiaAML/pull/24>`__
-   (`firefly-cpp <https://github.com/firefly-cpp>`__)
-
-`0.1.3 <https://github.com/lukapecnik/niaaml/tree/0.1.3>`__ (2020-12-04)
-------------------------------------------------------------------------
-
-`Full
[…]
