### 2021-01-09 #2949 @adi3
@lukapecnik thank you for your submission to JOSS. Your paper is a great start, but can do with a few modifications. Here are a few recommendations:

1. The summary section gets a little repetitive. For instance:
- "The framework incorporates stochastic population-based nature-inspired algorithms to search for the best classification pipeline"
- "NiaAML is slightly different, regarding the search for a perfect pipeline with the use of stochastic population-based nature-inspired algorithms"
- "In conclusion, NiaAML is an AutoML framework based on stochastic population-based nature-inspired algorithms."

Please edit the summary to avoid such repetitions.

2. The first paragraph of your Summary should to belong to the Statement of Need section.

3. I would like to see some content, if only a couple of lines, on what this "nature-inspired algorithm" is. How is it nature-inspired and why is that a good approach? It seems like that is the critical scholarly effort here, and without any explanation around this concept, there is nothing that clearly distinguishes NiaAML from other AutoML libraries.
