+++ paper/paper.bib
@@ -1,3 +1,10 @@
+@article{fister2013brief,
+  title={A brief review of nature-inspired algorithms for optimization},
+  author={Fister Jr, Iztok and Yang, Xin-She and Fister, Iztok and Brest, Janez and Fister, Du{\v{s}}an},
+  journal={arXiv preprint arXiv:1307.4186},
+  year={2013}
+}
+
 @incollection{Fister2020Continuous,
   title={Continuous Optimizers for Automatic Design and Evaluation of Classification Pipelines},
   author={Fister Jr., Iztok and Zorman, Milan and Fister, Du{\v{s}}an and Fister, Iztok},
@@ -42,13 +49,16 @@
   pages={1-8},
   doi={10.1109/IJCNN.2015.7280767}}
 
-@inproceedings{tpot,
-  title={TPOT: A tree-based pipeline optimization tool for automating machine learning},
-  author={Olson, Randal S and Moore, Jason H},
-  booktitle={Workshop on automatic machine learning},
-  pages={66--74},
-  year={2016},
-  organization={PMLR}
+
+@incollection{tpot,
+  doi = {10.1007/978-3-030-05318-5_8},
+  url = {https://doi.org/10.1007/978-3-030-05318-5_8},
+  year = {2019},
+  publisher = {Springer International Publishing},
+  pages = {151--160},
+  author = {Randal S. Olson and Jason H. Moore},
+  title = {{TPOT}: A Tree-Based Pipeline Optimization Tool for Automating Machine Learning},
+  booktitle = {Automated Machine Learning}
 }
 
 @book{bookHutter,
@@ -65,4 +75,32 @@
   timestamp = {Fri, 27 Mar 2020 08:32:01 +0100},
   biburl    = {https://dblp.org/rec/books/sp/HKV2019.bib},
   bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+
+@book{engelbrecht2007computational,
+  title={Computational intelligence: an introduction},
+  author={Engelbrecht, Andries P},
+  year={2007},
+  publisher={John Wiley \& Sons},
+  doi={10.1002/9780470512517}
+}
+
+@incollection{yang2014,
+  doi = {10.1016/b978-0-12-416743-8.00017-8},
+  url = {https://doi.org/10.1016/b978-0-12-416743-8.00017-8},
+  year = {2014},
+  publisher = {Elsevier},
+  title = {Nature-Inspired Optimization Algorithms},
+  booktitle = {Nature-Inspired Optimization Algorithms}
+  author={Yang, Xin-She},
+}
+
+@article{tzanetos2020comprehensive,
+  title={A comprehensive database of Nature-Inspired Algorithms},
+  author={Tzanetos, Alexandros and Fister Jr, Iztok and Dounias, Georgios},
+  journal={Data in Brief},
+  pages={105792},
+  year={2020},
+  publisher={Elsevier}
+  doi={10.1016/j.dib.2020.105792}
 }
