+++ paper/paper.bib
@@ -91,8 +91,8 @@
   year = {2014},
   publisher = {Elsevier},
   title = {Nature-Inspired Optimization Algorithms},
-  booktitle = {Nature-Inspired Optimization Algorithms}
-  author={Yang, Xin-She},
+  booktitle = {Nature-Inspired Optimization Algorithms},
+  author={Yang, Xin-She}
 }
 
 @article{tzanetos2020comprehensive,
@@ -101,6 +101,6 @@
   journal={Data in Brief},
   pages={105792},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/j.dib.2020.105792}
 }
