### 2021-01-12 42cc2ae
itemize paper.md fix [ci skip]
