### 2021-01-12 #2949 @lukapecnik
The authors would like to thank the Reviewer @adi3  sincerely for the
time spent on reading and commenting on the paper.

The given remarks are very thorough, and extremely helpful to overcome
the shortcomings and ambiguities in our paper.

In what follows, a point-by-point reply to the Reviewer’s remarks,
comments and suggestions is given.

--------------------------------------------------------------------

1. Repetitive sentences are now omitted from the manuscript.

2. As per your request, the first paragraph in the Summary Section has
been moved to the Statement of Need Section.

3. Good point! We have added a few lines of nature-inspired algorithms`
description. Moreover, we have also added some
literature about nature-inspired algorithms where the readers can gain
more information about these concepts.
Arguments that distinguish NiaAML from the other similar solutions in a
little bit more detail have now been added to
our paper.
