+++ paper/paper.md
@@ -26,7 +26,7 @@ The growing interest and development of Machine Learning (ML) methods has led to
 
 AutoML can be modeled as a continuous optimization problem with several potential optimization methods applied. Stochastic population-based nature-inspired algorithms [@yang2014; @engelbrecht2007computational] are a very popular tool for dealing with such continuous optimization problems. These algorithms are inspired mainly by the biological principles and phenomena of the behavior of various species living in nature [@fister2013brief]. Each stochastic population-based nature-inspired algorithm is composed of a population of individuals that undergo different variation operators during the evolution process that results in new populations. We developed the AutoML framework NiaAML, written fully in the Python programming language. NiaAML is based mainly on the method initially proposed in [@Fister2020Continuous]. The framework incorporates stochastic population-based nature-inspired algorithms to search for the best classification pipeline [@Fister2020Continuous].
 
-The framework is developed in a layer style layout architecture, consisting of several components, i.e. Feature Selection algorithms, Feature Transformation algorithms and classifiers. Its task is to find a perfect combination of components with proper classifiers hyperparameters` settings to build an efficient, yet customizable classification pipeline with the help of a popular collection of nature-inspired algorithms, named NiaPy [@Vrbančič2018]. Two types of optimizations in the NiaAML follow: (1) The first to find the optimal set of components for the pipeline, and (2) The second to tune the hyperparameters. Users can choose the ML components to be included into the optimization process freely, as well as select suitable fitness functions to be used for evaluation of candidate pipelines. Input data can be in the form of numerical and categorical features, as well as missing attributes, while pipelines are exported and imported as binary files for post-hoc use. Further, they can be exported as user-friendly text files that contain all of the relevant information about the pipeline and its components. A graphical outline of the NiaAML method is presented in Fig.\autoref{fig:NiaAMLflow}.
+The framework is developed in a layer style layout architecture, consisting of several components, i.e. Feature Selection algorithms, Feature Transformation algorithms and classifiers. Its task is to find a perfect combination of components with proper classifiers hyperparameters` settings to build an efficient, yet customizable classification pipeline with the help of a popular collection of nature-inspired algorithms, named NiaPy [@Vrbančič2018]. Two types of optimizations in the NiaAML follow: (1) The first to find the optimal set of components for the pipeline, and (2) The second to tune the hyperparameters. Users can choose the ML components to be included into the optimization process freely, as well as select suitable fitness functions to be used for evaluation of candidate pipelines. Input data can be in the form of numerical and categorical features, as well as missing attributes, while pipelines are exported and imported as binary files for post-hoc use. Further, they can be exported as user-friendly text files that contain all of the relevant information about the pipeline and its components. A graphical outline of the NiaAML method is presented in \autoref{fig:NiaAMLflow}.
 
 ![NiaAML flow.\label{fig:NiaAMLflow}](niaamlFlow.png)
 
@@ -40,12 +40,14 @@ NiaAML is different: Compared to similar Python AutoML frameworks, such as TPOT 
 
 - every pipeline in the output is feasible and functional in cases of the correct specified hyperparameters' domains,
 
-- the included GUI simplifies the work for regular users.
+- a Graphical User Interface (GUI) simplifies the work for users without expert domain knowledge.
 
-In conclusion, NiaAML is an AutoML framework based on stochastic population-based nature-inspired algorithms. Thus, the NiaAML framework is able to find the optimal classification pipelines built on preprocessing steps and classifiers, and, simultaneously, assuring a user-friendly experience. Working with ML is also, with the help of the NiaPy framework, more easy, adaptable, expandable and customizable. Last but not least, we can assure that more of the ML components and functionalities to the NiaAML framework are yet to come.
[…]
