### 2021-01-14 #2949 @adi3
@lukapecnik Thanks for the changes. I will be making some editorial changes myself to the draft and issuing a pull request. Please review it and merge it into your master branch. These edits will mostly be to correct syntactic usage or word choices. The aim is to keep the article crisp and to the point. It is faster for me to make these changes and issue a pull request instead of leading you through them. Let me know if you have any questions. Once you have merged the pull request, please notify me here so we can proceed.
