### 2021-01-14 #2949 @adi3
@lukapecnik Thank you. While I go through the remainder of the checklist, can you please figure out why one of your references is showing up as having a missing DOI?

```
MISSING DOIs

- 10.1007/978-3-030-05318-5_8 may be a valid DOI for title: TPOT: A tree-based pipeline optimization tool for automating machine learning
```
