### 2021-01-27 #2949 @adi3
@lukapecnik I'm reviewing your installation instructions.

- I recommend you specify supported OS that your software has been tested on and is confirmed to work. For instance, I was able to do pip installation just fine on Ubuntu, but ran into hurdles installing it on RHEL.
- The conda installation is failing at my end. I'm opening an issue on your repo so you can debug and look into it.

Please let me know when these two points have been fixed and I'll proceed. Thanks!
