### 2021-01-31 #2949 @lukapecnik
@adi3 Thank you.

- We have specified the OS we have tested the NiaAML package on.
- Good point, we have switched pip to pip3 as you suggested.
- Thank you for pointing out the `conda` failure. We surely will fix this option in the near future, but the investigation may take some time as after a few days we still have not found the reason why it is not working. That is why we decided to temporarily disable the installation using `conda` so the process of reviewing the package may proceed for now if you agree.

Thank you.
