### 2021-03-08 c1af416
example added, feature type checking updated
