+++ docs/changelog.rst
@@ -1,17 +1,37 @@
 Changelog
 =========
 
-`1.1.1rc2 <https://github.com/lukapecnik/niaaml/tree/1.1.1rc2>`__ (2020-12-22)
+`1.1.1 <https://github.com/lukapecnik/NiaAML/tree/1.1.1>`__ (2021-03-09)
+------------------------------------------------------------------------
+
+`Full
+Changelog <https://github.com/lukapecnik/NiaAML/compare/1.1.1rc2...1.1.1>`__
+
+**Merged pull requests:**
+
+-  logo [ci skip] `#48 <https://github.com/lukapecnik/NiaAML/pull/48>`__
+   (`lukapecnik <https://github.com/lukapecnik>`__)
+-  Fedora package pushed to stable
+   `#46 <https://github.com/lukapecnik/NiaAML/pull/46>`__
+   (`firefly-cpp <https://github.com/firefly-cpp>`__)
+-  Restructuring of paper and editorial changes
+   `#45 <https://github.com/lukapecnik/NiaAML/pull/45>`__
+   (`adi3 <https://github.com/adi3>`__)
+-  Corrections of paper
+   `#44 <https://github.com/lukapecnik/NiaAML/pull/44>`__
+   (`firefly-cpp <https://github.com/firefly-cpp>`__)
+
+`1.1.1rc2 <https://github.com/lukapecnik/NiaAML/tree/1.1.1rc2>`__ (2020-12-22)
 ------------------------------------------------------------------------------
 
 `Full
-Changelog <https://github.com/lukapecnik/niaaml/compare/1.1.1rc1...1.1.1rc2>`__
+Changelog <https://github.com/lukapecnik/NiaAML/compare/1.1.1rc1...1.1.1rc2>`__
 
-`1.1.1rc1 <https://github.com/lukapecnik/niaaml/tree/1.1.1rc1>`__ (2020-12-22)
+`1.1.1rc1 <https://github.com/lukapecnik/NiaAML/tree/1.1.1rc1>`__ (2020-12-22)
 ------------------------------------------------------------------------------
 
 `Full
-Changelog <https://github.com/lukapecnik/niaaml/compare/1.1.0...1.1.1rc1>`__
+Changelog <https://github.com/lukapecnik/NiaAML/compare/1.1.0...1.1.1rc1>`__
 
 **Merged pull requests:**
 
@@ -19,17 +39,17 @@ Changelog <https://github.com/lukapecnik/niaaml/compare/1.1.0...1.1.1rc1>`__
    `#43 <https://github.com/lukapecnik/NiaAML/pull/43>`__
    (`lukapecnik <https://github.com/lukapecnik>`__)
 
-`1.1.0 <https://github.com/lukapecnik/niaaml/tree/1.1.0>`__ (2020-12-16)
+`1.1.0 <https://github.com/lukapecnik/NiaAML/tree/1.1.0>`__ (2020-12-16)
 ------------------------------------------------------------------------
 
 `Full
-Changelog <https://github.com/lukapecnik/niaaml/compare/1.0.0rc7...1.1.0>`__
+Changelog <https://github.com/lukapecnik/NiaAML/compare/1.0.0rc7...1.1.0>`__
 
-`1.0.0rc7 <https://github.com/lukapecnik/niaaml/tree/1.0.0rc7>`__ (2020-12-14)
+`1.0.0rc7 <https://github.com/lukapecnik/NiaAML/tree/1.0.0rc7>`__ (2020-12-14)
 ------------------------------------------------------------------------------
 
 `Full
[…]
