+++ CHANGELOG.md
@@ -1,24 +1,35 @@
 # Changelog
 
-## [1.1.1rc2](https://github.com/lukapecnik/niaaml/tree/1.1.1rc2) (2020-12-22)
+## [1.1.1](https://github.com/lukapecnik/NiaAML/tree/1.1.1) (2021-03-09)
 
-[Full Changelog](https://github.com/lukapecnik/niaaml/compare/1.1.1rc1...1.1.1rc2)
+[Full Changelog](https://github.com/lukapecnik/NiaAML/compare/1.1.1rc2...1.1.1)
 
-## [1.1.1rc1](https://github.com/lukapecnik/niaaml/tree/1.1.1rc1) (2020-12-22)
+**Merged pull requests:**
 
-[Full Changelog](https://github.com/lukapecnik/niaaml/compare/1.1.0...1.1.1rc1)
+- logo \[ci skip\] [\#48](https://github.com/lukapecnik/NiaAML/pull/48) ([lukapecnik](https://github.com/lukapecnik))
+- Fedora package pushed to stable [\#46](https://github.com/lukapecnik/NiaAML/pull/46) ([firefly-cpp](https://github.com/firefly-cpp))
+- Restructuring of paper and editorial changes [\#45](https://github.com/lukapecnik/NiaAML/pull/45) ([adi3](https://github.com/adi3))
+- Corrections of paper [\#44](https://github.com/lukapecnik/NiaAML/pull/44) ([firefly-cpp](https://github.com/firefly-cpp))
+
+## [1.1.1rc2](https://github.com/lukapecnik/NiaAML/tree/1.1.1rc2) (2020-12-22)
+
+[Full Changelog](https://github.com/lukapecnik/NiaAML/compare/1.1.1rc1...1.1.1rc2)
+
+## [1.1.1rc1](https://github.com/lukapecnik/NiaAML/tree/1.1.1rc1) (2020-12-22)
+
+[Full Changelog](https://github.com/lukapecnik/NiaAML/compare/1.1.0...1.1.1rc1)
 
 **Merged pull requests:**
 
 - Original NiaAML method support \[ci skip\] [\#43](https://github.com/lukapecnik/NiaAML/pull/43) ([lukapecnik](https://github.com/lukapecnik))
 
-## [1.1.0](https://github.com/lukapecnik/niaaml/tree/1.1.0) (2020-12-16)
+## [1.1.0](https://github.com/lukapecnik/NiaAML/tree/1.1.0) (2020-12-16)
 
-[Full Changelog](https://github.com/lukapecnik/niaaml/compare/1.0.0rc7...1.1.0)
+[Full Changelog](https://github.com/lukapecnik/NiaAML/compare/1.0.0rc7...1.1.0)
 
-## [1.0.0rc7](https://github.com/lukapecnik/niaaml/tree/1.0.0rc7) (2020-12-14)
+## [1.0.0rc7](https://github.com/lukapecnik/NiaAML/tree/1.0.0rc7) (2020-12-14)
 
-[Full Changelog](https://github.com/lukapecnik/niaaml/compare/1.0.0rc6...1.0.0rc7)
+[Full Changelog](https://github.com/lukapecnik/NiaAML/compare/1.0.0rc6...1.0.0rc7)
 
 **Closed issues:**
 
@@ -29,9 +40,9 @@
 - Paper update [\#42](https://github.com/lukapecnik/NiaAML/pull/42) ([firefly-cpp](https://github.com/firefly-cpp))
 - minor corrections in examples [\#39](https://github.com/lukapecnik/NiaAML/pull/39) ([firefly-cpp](https://github.com/firefly-cpp))
 
-## [1.0.0rc6](https://github.com/lukapecnik/niaaml/tree/1.0.0rc6) (2020-12-12)
[…]
