### 2021-03-18 #2949 @lukapecnik
Hello @adi3 and @sara-02, I hope you are both doing well. How are things going forward with this review? Is there anything we can do so far to improve this framework and make it more suitable for publishing the paper in JOSS?

We are looking forward to hearing from you.
Thank you in advance.
