### 2021-03-18 #2949 @adi3
Hi @lukapecnik - I could not find any automated tests in your repo. Can you add some for the key functionalities, or if you already have them, document them in the readme?
