### 2021-03-18 #2949 @lukapecnik
Hi @adi3, the tests are available [here](https://github.com/lukapecnik/NiaAML/tree/master/niaaml/tests), and the instructions on how to run them are available in the `CONTRIBUTING.md`. They are also automated using the TravisCI platform (see `.travis.yml`). Test coverage is also available as a badge on top of the root `README.md` or via the Coveralls platform.

Do you think the instructions should be moved from the `CONTRIBUTING.md` to the root `README.md`?

Best regards
