### 2021-03-18 #2949 @adi3
I think leaving them in `CONTRIBUTING.md` is fine. Thanks for pointing them out to me. I'll run them and get back to you with feedback.
