### 2021-03-19 #2949 @adi3
@lukapecnik in your examples folder, can you please changes all occurrences of `python` to `python3`? This at least has to be the case in `run_all.sh` and `run_all.bat` files. Please do a search on other files as well. Since you are documenting the installation with `python3`, your examples should assume to run in that environment as well.
