### 2021-03-19 #2949 @adi3
@lukapecnik Can you please move the `Support` section from `CONTRIBUTING.md` to the main `README.md`? Also, please add a section for `Issues` in `README.md` and copy here the text you have under `Reporting Bugs` in `CONTRIBUTING.md`. Thanks! This is coming together pretty well.
