### 2021-03-20 #2949 @lukapecnik
Hi @adi3, thank you very much for your effort and quick response! I think I've improved the NiaAML repository now according to your great suggestions.

Best regards
