### 2021-03-21 #2949 @adi3
@lukapecnik thank you for making the changes. My checklist is complete and I will recommend this for publication. Good job!

@arfon I have completed my review against the JOSS checklist, and this submission meets our standard for publication.
