### 2021-03-28 #2949 @lukapecnik
Hello @arfon and @sara-02, we were wondering whether there has been some status update about this pending review. What are the next steps towards publishing? The authors of the NiaAML would be very happy if we can improve NiaAML with some further guidance and finish publishing in the following days. Thank you.
