### 2021-04-09 #2949 @lukapecnik
Hello @arfon and @sara-02! I hope you are well. We were wondering whether this review is going to be completed shortly. The authors would really like to use the JOSS paper as the base for further research in this research area.

We are looking forward to hearing from you soon.

Best regards
