### 2021-05-18 #2949 @sara-02
Hey @lukapecnik I was trying to run the[ following example](https://github.com/lukapecnik/NiaAML#example-of-usage) post installing `niaaml` and `Niapy` but when I copy pasted the example code to a python file and ran it, I am getting this error:
```
Traceback (most recent call last):
  File "niml.py", line 1, in <module>
    from niaaml import PipelineOptimizer, Pipeline
  File "/home/sarah/.local/lib/python3.8/site-packages/niaaml/__init__.py", line 3, in <module>
    from niaaml import preprocessing
  File "/home/sarah/.local/lib/python3.8/site-packages/niaaml/preprocessing/__init__.py", line 2, in <module>
    from niaaml.preprocessing import feature_selection
  File "/home/sarah/.local/lib/python3.8/site-packages/niaaml/preprocessing/feature_selection/__init__.py", line 3, in <module>
    from niaaml.preprocessing.feature_selection.jDEFSTH import jDEFSTH
  File "/home/sarah/.local/lib/python3.8/site-packages/niaaml/preprocessing/feature_selection/jDEFSTH.py", line 2, in <module>
    from NiaPy.algorithms.modified import SelfAdaptiveDifferentialEvolution
ModuleNotFoundError: No module named 'NiaPy'
```
