### 2021-05-19 a829657
Update NiaPy dependency, Add developer tools
