+++ niaaml/__init__.py
@@ -13,20 +13,20 @@ from niaaml.pipeline_component import PipelineComponent
 from niaaml.logger import Logger
 
 __all__ = [
-    'classifiers',
-    'data',
-    'preprocessing',
-    'fitness',
-    'get_bin_index',
-    'MinMax',
-    'ParameterDefinition',
-    'OptimizationStats',
-    'Factory',
-    'PipelineOptimizer',
-    'Pipeline',
-    'PipelineComponent',
-    'Logger'
+    "classifiers",
+    "data",
+    "preprocessing",
+    "fitness",
+    "get_bin_index",
+    "MinMax",
+    "ParameterDefinition",
+    "OptimizationStats",
+    "Factory",
+    "PipelineOptimizer",
+    "Pipeline",
+    "PipelineComponent",
+    "Logger",
 ]
 
-__project__ = 'niaaml'
-__version__ = '1.1.1'
+__project__ = "niaaml"
+__version__ = "1.1.2"
