### 2021-05-19 8deb97b
Update NiaPy dependency, Add developer tools
