+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
-## [1.1.1](https://github.com/lukapecnik/NiaAML/tree/1.1.1) (2021-03-09)
+## [1.1.2](https://github.com/lukapecnik/niaaml/tree/1.1.2) (2021-05-19)
 
-[Full Changelog](https://github.com/lukapecnik/NiaAML/compare/1.1.1rc2...1.1.1)
+[Full Changelog](https://github.com/lukapecnik/niaaml/compare/1.1.1...1.1.2)
+
+**Merged pull requests:**
+
+- Update niapy dependency [\#49](https://github.com/lukapecnik/NiaAML/pull/49) ([lukapecnik](https://github.com/lukapecnik))
+
+## [1.1.1](https://github.com/lukapecnik/niaaml/tree/1.1.1) (2021-03-09)
+
+[Full Changelog](https://github.com/lukapecnik/niaaml/compare/1.1.1rc2...1.1.1)
 
 **Merged pull requests:**
 
@@ -11,25 +19,25 @@
 - Restructuring of paper and editorial changes [\#45](https://github.com/lukapecnik/NiaAML/pull/45) ([adi3](https://github.com/adi3))
 - Corrections of paper [\#44](https://github.com/lukapecnik/NiaAML/pull/44) ([firefly-cpp](https://github.com/firefly-cpp))
 
-## [1.1.1rc2](https://github.com/lukapecnik/NiaAML/tree/1.1.1rc2) (2020-12-22)
+## [1.1.1rc2](https://github.com/lukapecnik/niaaml/tree/1.1.1rc2) (2020-12-22)
 
-[Full Changelog](https://github.com/lukapecnik/NiaAML/compare/1.1.1rc1...1.1.1rc2)
+[Full Changelog](https://github.com/lukapecnik/niaaml/compare/1.1.1rc1...1.1.1rc2)
 
-## [1.1.1rc1](https://github.com/lukapecnik/NiaAML/tree/1.1.1rc1) (2020-12-22)
+## [1.1.1rc1](https://github.com/lukapecnik/niaaml/tree/1.1.1rc1) (2020-12-22)
 
-[Full Changelog](https://github.com/lukapecnik/NiaAML/compare/1.1.0...1.1.1rc1)
+[Full Changelog](https://github.com/lukapecnik/niaaml/compare/1.1.0...1.1.1rc1)
 
 **Merged pull requests:**
 
 - Original NiaAML method support \[ci skip\] [\#43](https://github.com/lukapecnik/NiaAML/pull/43) ([lukapecnik](https://github.com/lukapecnik))
 
-## [1.1.0](https://github.com/lukapecnik/NiaAML/tree/1.1.0) (2020-12-16)
+## [1.1.0](https://github.com/lukapecnik/niaaml/tree/1.1.0) (2020-12-16)
 
-[Full Changelog](https://github.com/lukapecnik/NiaAML/compare/1.0.0rc7...1.1.0)
+[Full Changelog](https://github.com/lukapecnik/niaaml/compare/1.0.0rc7...1.1.0)
 
-## [1.0.0rc7](https://github.com/lukapecnik/NiaAML/tree/1.0.0rc7) (2020-12-14)
+## [1.0.0rc7](https://github.com/lukapecnik/niaaml/tree/1.0.0rc7) (2020-12-14)
 
-[Full Changelog](https://github.com/lukapecnik/NiaAML/compare/1.0.0rc6...1.0.0rc7)
+[Full Changelog](https://github.com/lukapecnik/niaaml/compare/1.0.0rc6...1.0.0rc7)
 
 **Closed issues:**
 
@@ -40,9 +48,9 @@
[…]
