### 2021-05-19 #2949 @lukapecnik
Hi @sara-02! Thank you for issue report. It seems that there's been a bit of an API change for the `niapy` package in the meantime. I created a new release of the `niaaml` (version 1.1.2) that should fix this problem. Please update the `niaaml` package and then the example should run smoothly.

Best regards
