+++ niaaml/classifiers/k_neighbors.py
@@ -1,8 +1,6 @@
 from niaaml.classifiers.classifier import Classifier
-from niaaml.utilities import MinMax
 from niaaml.utilities import ParameterDefinition
 from sklearn.neighbors import KNeighborsClassifier as KNC
-import numpy as np
 
 import warnings
 from sklearn.exceptions import (
