+++ niaaml/preprocessing/feature_transform/__init__.py
@@ -19,4 +19,5 @@ __all__ = [
     "MaxAbsScaler",
     "RobustScaler",
     "FeatureTransformAlgorithmFactory",
+    "QuantileTransformer",
 ]
