+++ niaaml/classifiers/gaussian_naive_bayes.py
@@ -1,8 +1,5 @@
 from niaaml.classifiers.classifier import Classifier
-from niaaml.utilities import MinMax
-from niaaml.utilities import ParameterDefinition
 from sklearn.naive_bayes import GaussianNB as GNB
-import numpy as np
 
 import warnings
 from sklearn.exceptions import (
