### 2021-05-23 #2949 @lukapecnik
Hello @arfon! I've archived the project using Zenodo. The DOI is `10.5281/zenodo.4782445` and the link to the archive is [https://zenodo.org/record/4782445#.YKqPs6gzaUl](url). I hope this is what you had in mind with updating this thread, otherwise please let me know.
